includes required updating as a result of moving tests/ int lib/

+++ lib/tests/xpath_result_set_test.cpp
@@ -1,6 +1,6 @@
-#include "xpath_result_set.hpp"
-#include "xpath_query.hpp"
-#include "xpath_ctxt.hpp"
+#include "lib/xpath_result_set.hpp"
+#include "lib/xpath_query.hpp"
+#include "lib/xpath_ctxt.hpp"
 
 #include "gtest/gtest.h"
 #include <limits.h>
@@ -8,6 +8,7 @@
 #include "test_helper.hpp"
 
 #include <iostream>
+#include <cstddef>
 
 class XPathResultSetTest : public testing::Test {
 protected:
@@ -20,7 +21,7 @@ private:
 
 TEST_F( XPathResultSetTest, DefaultCtor ) {
     XPathResultSet resultSet;
-    EXPECT_EQ( 0, resultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>( resultSet.size( ) ) );
 }
 
 TEST_F( XPathResultSetTest, Ctor ) {
@@ -28,17 +29,17 @@ TEST_F( XPathResultSetTest, Ctor ) {
     XPathCtxt nullXpc{nullXml};
     XPathQuery nullQuery{nullXpc};
     XPathResultSet nullResultSet{nullQuery};
-    EXPECT_EQ( 0, nullResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>( nullResultSet.size( ) ) );
 
     XPathQuery menuQuery{queryMenu( )};
     menuQuery.query( "//calories" );
     XPathResultSet menuResultSet{menuQuery};
-    EXPECT_EQ( 5, menuResultSet.size( ) );
+    EXPECT_EQ( 5, static_cast<const int>( menuResultSet.size( ) ) );
 
     XPathQuery booksQuery{queryBooks( )};
     booksQuery.query( "//bad/query" );
     XPathResultSet booksResultSet{booksQuery};
-    EXPECT_EQ( 0, booksResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(booksResultSet.size( ) ));
 }
 
 TEST_F( XPathResultSetTest, CopyCtor ) {
@@ -46,52 +47,52 @@ TEST_F( XPathResultSetTest, CopyCtor ) {
     XPathCtxt nullXpc{nullXml};
     XPathQuery nullQuery{nullXpc};
     XPathResultSet nullResultSet{nullQuery};
-    ASSERT_EQ( 0, nullResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     XPathResultSet cpNullResultSet{nullResultSet};
-    EXPECT_EQ( 0, cpNullResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(cpNullResultSet.size( ) ));
 
     XPathQuery menuQuery{queryMenu( )};
     menuQuery.query( "//calories" );
     XPathResultSet menuResultSet{menuQuery};
-    ASSERT_EQ( 5, menuResultSet.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathResultSet cpMenuResultSet{menuResultSet};
-    EXPECT_EQ( 5, cpMenuResultSet.size( ) );
+    EXPECT_EQ( 5, static_cast<const int>(cpMenuResultSet.size( ) ));
 
     XPathQuery booksQuery{queryBooks( )};
     booksQuery.query( "//bad/query" );
     XPathResultSet booksResultSet{booksQuery};
-    ASSERT_EQ( 0, booksResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(booksResultSet.size( ) ));
 
     XPathResultSet cpBooksResultSet{booksResultSet};
-    EXPECT_EQ( 0, cpBooksResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(cpBooksResultSet.size( ) ));
 }
 
 TEST_F( XPathResultSetTest, CopyMoveCtor ) {
     XPathCtxt nullXpc{XmlDoc{}};
     XPathQuery nullQuery{nullXpc};
     XPathResultSet nullResultSet{nullQuery};
-    ASSERT_EQ( 0, nullResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     XPathResultSet mvCpNullResultSet{cp_f<XPathResultSet>( nullResultSet )};
-    EXPECT_EQ( 0, mvCpNullResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(mvCpNullResultSet.size( ) ));
 
     XPathQuery menuQuery{queryMenu( )};
     menuQuery.query( "//calories" );
     XPathResultSet menuResultSet{menuQuery};
-    ASSERT_EQ( 5, menuResultSet.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathResultSet mvCpMenuResultSet{cp_f<XPathResultSet>( menuResultSet )};
-    EXPECT_EQ( 5, mvCpMenuResultSet.size( ) );
+    EXPECT_EQ( 5, static_cast<const int>(mvCpMenuResultSet.size( ) ));
 
     XPathQuery booksQuery{queryBooks( )};
     booksQuery.query( "//bad/query" );
     XPathResultSet booksResultSet{booksQuery};
-    ASSERT_EQ( 0, booksResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(booksResultSet.size( ) ));
 
     XPathResultSet mvCpBooksResultSet{cp_f<XPathResultSet>( booksResultSet )};
-    EXPECT_EQ( 0, mvCpBooksResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(mvCpBooksResultSet.size( ) ));
 }
 
 TEST_F( XPathResultSetTest, CopyAssignment ) {
@@ -100,62 +101,62 @@ TEST_F( XPathResultSetTest, CopyAssignment ) {
     XPathCtxt nullXpc{nullXml};
     XPathQuery nullQuery{nullXpc};
     XPathResultSet nullResultSet{nullQuery};
-    ASSERT_EQ( 0, nullResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     XPathQuery defaultQuery;
     XPathResultSet nullResultSet1{defaultQuery};
-    ASSERT_EQ( 0, nullResultSet1.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet1.size( ) ));
 
     nullResultSet = nullResultSet1;
-    ASSERT_EQ( 0, nullResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     // assign to null
     XPathQuery menuQuery{queryMenu( )};
     menuQuery.query( "//calories" );
     XPathResultSet menuResultSet{menuQuery};
-    ASSERT_EQ( 5, menuResultSet.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathResultSet cpNullResultSet1{nullResultSet};
-    ASSERT_EQ( 0, cpNullResultSet1.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(cpNullResultSet1.size( ) ));
 
     cpNullResultSet1 = menuResultSet;
-    EXPECT_EQ( 5, cpNullResultSet1.size( ) );
-    EXPECT_EQ( 5, menuResultSet.size( ) );
+    EXPECT_EQ( 5, static_cast<const int>(cpNullResultSet1.size( ) ));
+    EXPECT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathQuery booksQuery{queryBooks( )};
     booksQuery.query( "//bad/query" );
     XPathResultSet booksResultSet{booksQuery};
-    ASSERT_EQ( 0, booksResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(booksResultSet.size( ) ));
 
     XPathResultSet cpNullResultSet2{nullResultSet};
-    ASSERT_EQ( 0, cpNullResultSet2.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(cpNullResultSet2.size( ) ));
 
     cpNullResultSet2 = booksResultSet;
-    EXPECT_EQ( 0, cpNullResultSet2.size( ) );
-    EXPECT_EQ( 0, booksResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(cpNullResultSet2.size( ) ));
+    EXPECT_EQ( 0, static_cast<const int>(booksResultSet.size( ) ));
 
     // assign null
-    ASSERT_EQ( 5, cpNullResultSet1.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(cpNullResultSet1.size( ) ));
     cpNullResultSet1 = nullResultSet;
-    EXPECT_EQ( 0, cpNullResultSet1.size( ) );
-    EXPECT_EQ( 0, nullResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(cpNullResultSet1.size( ) ));
+    EXPECT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     // assign
     // booksQuery.query( "//title" );
     XPathQuery booksQuery1{queryBooks( )};
     booksQuery1.query( "//title" );
     XPathResultSet booksResultSet1{booksQuery1};
-    ASSERT_EQ( 12, booksResultSet1.size( ) );
-    ASSERT_EQ( 5, menuResultSet.size( ) );
+    ASSERT_EQ( 12, static_cast<const int>(booksResultSet1.size( ) ));
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathQuery menuQuery1{queryMenu( )};
     menuQuery1.query( "//calories" );
     XPathResultSet menuResultSet1{menuQuery1};
-    ASSERT_EQ( 5, menuResultSet1.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet1.size( ) ));
 
     menuResultSet1 = booksResultSet1;
-    EXPECT_EQ( 12, menuResultSet1.size( ) );
-    EXPECT_EQ( 12, booksResultSet1.size( ) );
+    EXPECT_EQ( 12, static_cast<const int>(menuResultSet1.size( ) ));
+    EXPECT_EQ( 12, static_cast<const int>(booksResultSet1.size( ) ));
 }
 
 TEST_F( XPathResultSetTest, CopyMoveAssignment ) {
@@ -164,61 +165,61 @@ TEST_F( XPathResultSetTest, CopyMoveAssignment ) {
     XPathCtxt nullXpc{nullXml};
     XPathQuery nullQuery{nullXpc};
     XPathResultSet nullResultSet{nullQuery};
-    ASSERT_EQ( 0, nullResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     XPathQuery defaultQuery;
     XPathResultSet nullResultSet1{defaultQuery};
-    ASSERT_EQ( 0, nullResultSet1.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet1.size( ) ));
 
     nullResultSet = cp_f<XPathResultSet>( nullResultSet1 );
-    ASSERT_EQ( 0, nullResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     // assign to null
     XPathQuery menuQuery{queryMenu( )};
     menuQuery.query( "//calories" );
     XPathResultSet menuResultSet{menuQuery};
-    ASSERT_EQ( 5, menuResultSet.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathResultSet cpNullResultSet1{nullResultSet};
-    ASSERT_EQ( 0, cpNullResultSet1.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(cpNullResultSet1.size( ) ));
 
     cpNullResultSet1 = cp_f<XPathResultSet>( menuResultSet );
-    EXPECT_EQ( 5, cpNullResultSet1.size( ) );
-    EXPECT_EQ( 5, menuResultSet.size( ) );
+    EXPECT_EQ( 5, static_cast<const int>(cpNullResultSet1.size( ) ));
+    EXPECT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathQuery booksQuery{queryBooks( )};
     booksQuery.query( "//bad/query" );
     XPathResultSet booksResultSet{booksQuery};
-    ASSERT_EQ( 0, booksResultSet.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(booksResultSet.size( ) ));
 
     XPathResultSet cpNullResultSet2{nullResultSet};
-    ASSERT_EQ( 0, cpNullResultSet2.size( ) );
+    ASSERT_EQ( 0, static_cast<const int>(cpNullResultSet2.size( ) ));
 
     cpNullResultSet2 = cp_f<XPathResultSet>( booksResultSet );
-    EXPECT_EQ( 0, cpNullResultSet2.size( ) );
-    EXPECT_EQ( 0, booksResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(cpNullResultSet2.size( ) ));
+    EXPECT_EQ( 0, static_cast<const int>(booksResultSet.size( ) ));
 
     // assign null
-    ASSERT_EQ( 5, cpNullResultSet1.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(cpNullResultSet1.size( ) ));
     cpNullResultSet1 = cp_f<XPathResultSet>( nullResultSet );
-    EXPECT_EQ( 0, cpNullResultSet1.size( ) );
-    EXPECT_EQ( 0, nullResultSet.size( ) );
+    EXPECT_EQ( 0, static_cast<const int>(cpNullResultSet1.size( ) ));
+    EXPECT_EQ( 0, static_cast<const int>(nullResultSet.size( ) ));
 
     // assign
     XPathQuery booksQuery1{queryBooks( )};
     booksQuery1.query( "//title" );
     XPathResultSet booksResultSet1{booksQuery1};
-    ASSERT_EQ( 12, booksResultSet1.size( ) );
-    ASSERT_EQ( 5, menuResultSet.size( ) );
+    ASSERT_EQ( 12, static_cast<const int>(booksResultSet1.size( ) ));
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet.size( ) ));
 
     XPathQuery menuQuery1{queryMenu( )};
     menuQuery1.query( "//calories" );
     XPathResultSet menuResultSet1{menuQuery1};
-    ASSERT_EQ( 5, menuResultSet1.size( ) );
+    ASSERT_EQ( 5, static_cast<const int>(menuResultSet1.size( ) ));
 
     menuResultSet1 = cp_f<XPathResultSet>( booksResultSet1 );
-    EXPECT_EQ( 12, menuResultSet1.size( ) );
-    EXPECT_EQ( 12, booksResultSet1.size( ) );
+    EXPECT_EQ( 12, static_cast<const int>(menuResultSet1.size( ) ));
+    EXPECT_EQ( 12, static_cast<const int>(booksResultSet1.size( ) ));
 }
 
 TEST_F( XPathResultSetTest, IteratorTest ) {
